requirements.txt
streamlit
opencv-python-headless
ezdxf
Pillow
numpy
cairosvg
matplotlib